alarm repoid 수정

See merge request s10-webmobile2-sub2/S10P12A709!193

+++ backend/coala/src/main/java/com/ssafy/coala/domain/alarm/dao/HelpAlarmRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface HelpAlarmRepository extends JpaRepository<HelpAlarm,Integer> {
+public interface HelpAlarmRepository extends JpaRepository<HelpAlarm,Long> {
 
     List<HelpAlarm> findByReceiver(String receiver);
